requirements.txt
-i https://pypi.org/simple
ansible==11.0.0; python_version >= '3.11'
ansible-core==2.18.0; python_version >= '3.11'
aptly-api-client==0.3.0
certifi==2024.8.30; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
cryptography==43.0.3; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
iso8601==2.1.0; python_version >= '3.7' and python_version < '4.0'
jinja2==3.1.4; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
markupsafe==3.0.2; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
pathspec==0.12.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
resolvelib==1.0.1
urllib3==2.2.3; python_version >= '3.8'
voluptuous==0.15.2; python_version >= '3.9'
